+++ app/data/local/room/dao/ProviderDao.kt
@@ -15,7 +15,7 @@ interface ProviderDao {
     fun getAllProviders(): Flow<List<ProviderEntity>>
 
     @Query("SELECT * FROM providers WHERE id = :providerId")
-    suspend fun getProviderById(providerId: Int): ProviderEntity?
+    suspend fun getProviderById(providerId: Long): ProviderEntity?
 
     @Insert(onConflict = OnConflictStrategy.IGNORE)
     suspend fun insertProvider(provider: ProviderEntity): Long
